+++ spark/common/src/main/scala/org/apache/sedona/spark/SedonaContext.scala
@@ -27,8 +27,8 @@ import org.apache.spark.serializer.KryoSerializer
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.execution.SparkStrategy
+import org.apache.spark.sql.sedona_sql.UDT.TransformNestedUDTInParquet
 import org.apache.spark.sql.sedona_sql.optimization._
-import org.apache.spark.sql.sedona_sql.catalyst.FixNestedUDTInParquetRule
 import org.apache.spark.sql.sedona_sql.strategy.join.JoinQueryDetector
 import org.apache.spark.sql.sedona_sql.strategy.physical.function.EvalPhysicalFunctionStrategy
 import org.apache.spark.sql.{SQLContext, SparkSession}
@@ -44,7 +44,7 @@ object SedonaContext {
 
   private def customOptimizationsWithSession(sparkSession: SparkSession) =
     Seq(
-      FixNestedUDTInParquetRule(sparkSession),
+      new TransformNestedUDTInParquet(sparkSession),
       new SpatialFilterPushDownForGeoParquet(sparkSession),
       new SpatialTemporalFilterPushDownForStacScan(sparkSession))
